requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyspark>=3.3.0
pyarrow>=10.0.0

# ML and NLP
torch>=1.12.0
sentence-transformers>=2.2.0
transformers>=4.21.0
faiss-cpu>=1.7.0
langchain>=0.0.200
langchain-community>=0.0.20

# API and Web
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
streamlit>=1.25.0
requests>=2.28.0

# Data processing
scikit-learn>=1.1.0
nltk>=3.8.0
spacy>=3.4.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
python-multipart>=0.0.5

# Testing and Quality
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=5.0.0
black>=22.0.0
isort>=5.10.0

# Monitoring
mlflow>=2.0.0
wandb>=0.13.0

# Development
pre-commit>=2.20.0
jupyter>=1.0.0
ipykernel>=6.15.0
